+++ schemacrawler-ai-core/src/test/java/schemacrawler/tools/ai/function/test/ListAcrossTablesFunctionTest.java
@@ -118,8 +118,6 @@ public class ListAcrossTablesFunctionTest extends AbstractFunctionTest {
         mapper.readValue(args.toString(), new TypeReference<Map<String, String>>() {});
 
     assertThat(resultMap, hasEntry("dependant_object_type", "NONE"));
-    assertThat(resultMap, hasEntry("dependant_object_name", ""));
-    assertThat(resultMap, hasEntry("table_name", null));
   }
 
   @Test